requirements.txt
opencv-python
opencv-contrib-python
pandas
Pillow
pymysql
openpyxl